+++ frontend/index.html
@@ -1,0 +1,9 @@
+<!DOCTYPE html>
+<html>  
+<head>
+	<title>Daft Punk</title>
+</head>
+<body>
+	<h1>Hello worlds</h1>
+</body>
+</html>
